Split domain package into two subpackages: modelmappers and usecases
