+++ slackhealthbot/scheduler.py
@@ -9,11 +9,11 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from slackhealthbot.core.exceptions import UserLoggedOutException
 from slackhealthbot.core.models import SleepData
 from slackhealthbot.database.connection import SessionLocal
-from slackhealthbot.domain.fitbit import (
+from slackhealthbot.domain.usecases.fitbit import (
     usecase_process_new_activity,
     usecase_process_new_sleep,
 )
-from slackhealthbot.domain.slack import usecase_post_user_logged_out
+from slackhealthbot.domain.usecases.slack import usecase_post_user_logged_out
 from slackhealthbot.repositories import fitbitrepository
 from slackhealthbot.repositories.fitbitrepository import UserIdentity
 from slackhealthbot.settings import settings
